Added callback methods for keypress
This change adds a keypress event handler for each of the sort headers. When the header is tabbed to and "enter" or "space" are pressed, the table will sort just as it would if the mouse was clicked. This makes this module more ADA compliant by allowing non-sighted users to sort the tables as well.